--- lib/chartmogul/import/transactions/payment.rb
@@ -1,26 +1,0 @@
-module ChartMogul
-  module Import
-    module Transactions
-      class Payment < APIResource
-        set_resource_name 'Payment Transaction'
-        set_resource_path '/v1/import/invoices/:invoice_uuid/transactions'
-
-        readonly_attr :uuid
-
-        writeable_attr :type, default: 'payment'
-        writeable_attr :date, type: :time
-        writeable_attr :result
-        writeable_attr :external_id
-
-        writeable_attr :invoice_uuid
-
-        def initialize(attributes = {})
-          super(attributes)
-          @type = 'payment'
-        end
-
-        include API::Actions::Create
-      end
-    end
-  end
-end
